Перенесёт фукнции для рендеринга из `utils` в  `utils/render.js`
