+++ src/utils.js
@@ -1,8 +1,3 @@
-// Перечисления места для вставки DOM-элемента
-const RenderPosition = {
-  AFTER_BEGIN: `afterbegin`,
-  BEFORE_END: `beforeend`,
-};
 // Получаем произвольный item массива
 const getRandomItem = (array) => { // изменить имя на getRandomItem
   const random = Math.floor(Math.random() * array.length);
@@ -25,24 +20,6 @@ const getWatchedCount = (movieCards) => {
 };
 const getFavoriteCount = (movieCards) => {
   return movieCards.filter((card) => card.isFavorite);
-};
-// Функция для создания DOM-элемента
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-// Функция для отрисовки созданного DOM-элемента
-const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTER_BEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFORE_END:
-      container.append(element);
-      break;
-  }
 };
 // Получаем максимальное значение рейтинга и комментариев для экстракарточек
 const getIndexRatingCards = (cards, propPath) => {
@@ -74,4 +51,4 @@ const getIndexRatingCards = (cards, propPath) => {
   };
 };
 
-export {getRandomItem, getRandomIndex, getRandomIntervalNumber, getWatchlistCount, getWatchedCount, getFavoriteCount, createElement, render, RenderPosition, getIndexRatingCards};
+export {getRandomItem, getRandomIndex, getRandomIntervalNumber, getWatchlistCount, getWatchedCount, getFavoriteCount, getIndexRatingCards};
