+++ src/utils/render.js
@@ -1,0 +1,25 @@
+// Перечисления места для вставки DOM-элемента
+const RenderPosition = {
+  AFTER_BEGIN: `afterbegin`,
+  BEFORE_END: `beforeend`,
+};
+// Функция для создания DOM-элемента
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+// Функция для отрисовки созданного DOM-элемента
+const render = (container, element, place) => {
+  switch (place) {
+    case RenderPosition.AFTER_BEGIN:
+      container.prepend(element);
+      break;
+    case RenderPosition.BEFORE_END:
+      container.append(element);
+      break;
+  }
+};
+
+export {createElement, render, RenderPosition};
